requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
pydantic==1.10.14
python-dotenv==1.0.1
email-validator==2.1.0
